I would like to add one new option, which is common to all CLIs. However, commonality of individual CLI construction is lacking and maintenance is difficult. For this reason, we introduced cli_common.py.
